fix code for deploy
