+++ project1/src/types/Blog.ts
@@ -9,5 +9,5 @@ export interface Blog {
   tags: number[];
   date: string;
   created_at: string;
-  author_id: number;
+  author_id: string;
 }
