+++ project1/src/types/User.ts
@@ -9,4 +9,5 @@ export interface User {
   role: 'user' | 'admin';
   createdAt?: string;
   updatedAt?: string;
+  website?: string;
 }
