+++ project1/src/types/UserForComment.ts
@@ -1,4 +1,5 @@
 export interface UserForComment {
   id: string;
   fullName: string;
+  image?: string;
 }
